# specs/001-lims-core/contracts/openapi.yaml
openapi: 3.2.0
info:
  title: Laboratory Information Management System API
  version: 0.1.0
  description: 高校科研实验室信息管理系统 API 契约初稿
servers:
  - url: /api/v1
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseLogin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /users:
    get:
      tags: [Admin]
      summary: 查询用户
      operationId: listUsers
      parameters:
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [ACTIVE, INACTIVE] }
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseAdminUsers' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin]
      summary: 创建用户
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminUserCreateRequest' }
      responses:
        '201':
          description: 用户创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseAdminUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /methods:
    get:
      tags: [Methods]
      summary: 查询实验方法版本
      operationId: listMethods
      parameters:
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string, enum: [DRAFT, ACTIVE, INACTIVE, EXPIRED] }
      responses:
        '200':
          description: 方法版本列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseMethodItems' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Methods]
      summary: 创建实验方法版本
      operationId: createMethod
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MethodCreateRequest' }
      responses:
        '201': { description: 创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponseMethod' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /methods/{id}/attachments:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Methods]
      summary: 上传并登记实验方法附件
      operationId: createMethodAttachment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MethodAttachmentRequest' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/MethodAttachmentUploadRequest' }
      responses:
        '201': { description: 登记成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponseMethodAttachment' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /tasks:
    get:
      tags: [Tasks]
      summary: 查询任务
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: projectId
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTaskPage'
    post:
      tags: [Tasks]
      summary: 创建实验任务
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseTask'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServiceUnavailable:
      description: 服务配置不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: 业务状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        traceId: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    LoginData:
      type: object
      required: [user]
      properties:
        user:
          type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email, nullable: true }
    AdminUserCreateRequest:
      type: object
      required: [email, password, username, realName]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        username: { type: string, maxLength: 64 }
        realName: { type: string, maxLength: 64 }
        roleCodes:
          type: array
          items: { type: string }
...
    AdminUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        realName: { type: string }
        email: { type: string, format: email, nullable: true }
        status: { type: string }
        roles:
          type: array
          items: { $ref: '#/components/schemas/AdminRole' }
...
    AdminRole:
      type: object
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        status: { type: string }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/AdminPermission' }
    AdminPermission:
      type: object
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        resource: { type: string }
        action: { type: string }
    ApiResponseAdminUser:
      type: object
      properties:
        data: { $ref: '#/components/schemas/AdminUser' }
    ApiResponseAdminUsers:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/AdminUser' }
...
    MethodCreateRequest:
      type: object
      required: [methodCode, name, version]
      properties:
        methodCode: { type: string, maxLength: 32 }
        name: { type: string, maxLength: 128 }
        version: { type: string, maxLength: 32 }
        scope: { type: string, maxLength: 255, nullable: true }
        detectionLimit: { type: number, minimum: 0, nullable: true }
        status: { type: string, enum: [DRAFT, ACTIVE, INACTIVE, EXPIRED] }
        documentId: { type: integer, format: int64, nullable: true }
        effectiveAt: { type: string, format: date-time, nullable: true }
        expiredAt: { type: string, format: date-time, nullable: true }
...
    MethodAttachmentRequest:
      type: object
      required: [fileName, storagePath, fileSize, contentType]
      properties:
        fileName: { type: string, maxLength: 255 }
        storagePath: { type: string, maxLength: 512 }
        fileSize: { type: integer, format: int64, minimum: 0 }
        contentType: { type: string, maxLength: 128 }
    MethodHistory:
      type: object
      properties:
        id: { type: integer, format: int64 }
        methodId: { type: integer, format: int64 }
        methodCode: { type: string }
        fromVersion: { type: string, nullable: true }
        toVersion: { type: string }
        fromStatus: { type: string, nullable: true }
        toStatus: { type: string }
        changeType: { type: string, enum: [CREATE_VERSION, UPDATE, STATUS_CHANGE] }
        operatorId: { type: string, format: uuid }
        remark: { type: string, nullable: true }
        occurredAt: { type: string, format: date-time }
    MethodAttachment:
      type: object
      properties:
        id: { type: integer, format: int64 }
        objectType: { type: string }
        objectId: { type: string }
        fileName: { type: string }
        storagePath: { type: string }
        fileSize: { type: integer, format: int64 }
        contentType: { type: string }
        uploadedBy: { type: string, format: uuid }
        uploadedAt: { type: string, format: date-time }
    MethodAttachmentUploadRequest:
      type: object
      required: [file]
      properties:
        file: { type: string, format: binary, description: 实验方法或相关标准规范文件 }
    Method:
      type: object
      properties:
        id: { type: integer, format: int64 }
        methodCode: { type: string }
        name: { type: string }
        version: { type: string }
        scope: { type: string, nullable: true }
        detectionLimit: { type: number, nullable: true }
        status: { type: string, enum: [DRAFT, ACTIVE, INACTIVE, EXPIRED] }
        documentId: { type: integer, format: int64, nullable: true }
        effectiveAt: { type: string, format: date-time, nullable: true }
        expiredAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        history: { type: array, items: { $ref: '#/components/schemas/MethodHistory' } }
        attachments: { type: array, items: { $ref: '#/components/schemas/MethodAttachment' } }
    ApiResponseMethod:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        data: { $ref: '#/components/schemas/Method' }
        traceId: { type: string }
    ApiResponseMethodItems:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        data: { type: array, items: { $ref: '#/components/schemas/Method' } }
        traceId: { type: string }
    ApiResponseMethodAttachment:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        data: { $ref: '#/components/schemas/MethodAttachment' }
        traceId: { type: string }
...
    TaskCreateRequest:
      type: object
      required: [taskCode, projectId, methodId, name, priority]
      properties:
        taskCode: { type: string, maxLength: 32 }
        projectId: { type: integer, format: int64 }
        methodId: { type: integer, format: int64 }
        name: { type: string }
        priority: { type: string, enum: [LOW, NORMAL, HIGH] }
        plannedStart: { type: string, format: date }
        plannedEnd: { type: string, format: date }
        remark: { type: string, nullable: true }
        sampleIds:
          type: array
          items: { type: integer, format: int64 }
...
    Task:
      type: object
      properties:
        id: { type: integer, format: int64 }
        taskCode: { type: string }
        projectId: { type: integer, format: int64 }
        methodId: { type: integer, format: int64 }
        name: { type: string }
        priority: { type: string }
        status: { type: string }
        plannedStart: { type: string, format: date, nullable: true }
        plannedEnd: { type: string, format: date, nullable: true }
        remark: { type: string, nullable: true }
        sampleIds: { type: array, items: { type: integer, format: int64 } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApiResponseLogin:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        data: { $ref: '#/components/schemas/LoginData' }
        traceId: { type: string }
...
    ApiResponseTask:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        data: { $ref: '#/components/schemas/Task' }
        traceId: { type: string }
    ApiResponseTaskPage:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Task' }
            page: { type: integer }
            size: { type: integer }
            total: { type: integer, format: int64 }
        traceId: { type: string }
...
security:
  - bearerAuth: []